RedisService Methods

+++ CarRental.API/Controllers/CarControllers/CarController.cs
@@ -4,8 +4,9 @@ using CarRental.Core.DTOs;
 using CarRental.Core.DTOs.CarDTOs;
 using CarRental.Core.Models;
 using CarRental.Core.Services;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json;
+using System.Text;
 
 namespace CarRental.API.Controllers.CarControllers
 {
@@ -13,18 +14,23 @@ namespace CarRental.API.Controllers.CarControllers
     {
         private readonly ICarService _service;
         private readonly IMapper _mapper;
+        private readonly RedisService _redisService;
 
 
         public CarController(ICarService service, IMapper mapper, RedisService redisService)
         {
             _service = service;
             _mapper = mapper;
+            _redisService = redisService;
+
         }
 
         [HttpGet]
         public async Task<IActionResult> All()
         {
+            
             var cars = await _service.GetAllAsync();
+          
             var carsDto = _mapper.Map<List<GetCarDto>>(cars);
             return CreateActionResult(CustomResponseDto<List<GetCarDto>>.Success(200, carsDto.ToList()));
 
@@ -35,6 +41,9 @@ namespace CarRental.API.Controllers.CarControllers
         public async Task<IActionResult> GetById(int id)
         {
             var car = await _service.GetByIdAsync(id);
+            if (car == null)
+                return CreateActionResult(CustomResponseDto<NoContentDto>.Fail(404, "Araba Bulunamadı!"));
+
             var carDto = _mapper.Map<GetCarDto>(car);
             return CreateActionResult(CustomResponseDto<GetCarDto>.Success(200, carDto));
         }
@@ -42,16 +51,26 @@ namespace CarRental.API.Controllers.CarControllers
         [HttpPost]
         public async Task<IActionResult> Save(SetCarDto carDto)
         {
-           
+            var cars = await _service.GetAllAsync();
+
             var car = _mapper.Map<Car>(carDto);
             car.IsRent = true;
+
+            foreach (var item in cars)
+            {
+                if (item.LicensePlate==car.LicensePlate)
+                {
+                    return CreateActionResult(CustomResponseDto<NoContentDto>.Fail(403, "Araç Sistemde Kayıtlı!"));
+                }
+            }
+
             await _service.AddAsync(car);
             return CreateActionResult(CustomResponseDto<SetCarDto>.Success(201, carDto));
 
         }
 
         [HttpPut]
-        public async Task<IActionResult> Update(SetCarDto carDto)   
+        public async Task<IActionResult> Update(SetCarDto carDto)
         {
             var car = _mapper.Map<Car>(carDto);
             await _service.UpdateAsync(car);
@@ -63,6 +82,10 @@ namespace CarRental.API.Controllers.CarControllers
         public async Task<IActionResult> Remove(int id)
         {
             var car = await _service.GetByIdAsync(id);
+            if (car==null)
+            {
+                return CreateActionResult(CustomResponseDto<NoContentDto>.Fail(404, "Araba Bulunamadı!"));
+            }
             await _service.RemoveAsync(car);
             return CreateActionResult(CustomResponseDto<NoContentDto>.Success(204));
